Refactor uart{1,2}.* into uart.*
Instead of having a bunch of duplicate code in UART1_*/UART2_*
functions, functionality is now collected into a single UART_*
function. UART1 or UART2 is selected via a `select` parameter to the
function, which shall have a value of either U1SELECT or U2SELECT.

For a few functions, UART1_* wrapper functions are provided for
convenience.