Remove broken modules
There are transitive and circular dependencies everywhere. This began with
removing src/helpers/interval.*, which is superceded by the new logic analyzer
implementation. But the old IC driver transitivelly depended on interval, so
it had to go. Then a bunch of things that depended on IC had to go, and so on
and so forth.

Basically, better to start fresh at this point.